requirements.txt
azure-cognitiveservices-speech>=1.31.0
azure-identity==1.15.0
openai>=1.0.0
python-dotenv==1.0.0
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.10.0
secure-smtplib==0.1.1
gunicorn==21.2.0    